requirements.txt
numpy==1.26.4
torch
torchvision
segmentation-models-pytorch
timm
opencv-python
pillow
lxml
PyEMD==1.0.0